chore : rename interfaces into typings
